+++ src/interpreter/mod.rs
@@ -682,3 +682,5 @@ pub fn run_func(func: FuncImpl, args: &[Expression], env: Rc<RefCell<Env>>) -> L
     }
     LiteralType::Null
 }
+
+// @todo better organized statements