+++ src/resolver.rs
@@ -332,3 +332,5 @@ impl Resolver {
         }
     }
 }
+
+// @todo type locals for generics