+++ src/std/mod.rs
@@ -677,3 +677,5 @@ impl Interpreter {
         }
     }
 }
+
+// @todo enlarge the std and introduce better managing system